Namespacing
Huh. Right: if the function is called by the JQuery `$().each()` method, then the object that `this` refers to must be the object on which `each()` was called, not the original namespace of the function being called. So... easier to just redefine the dang array of colors every time.

Also: let's just push out a default color using Canvas course colors API (that allows folks to customize the non-color blocks, for one thing, and it makes the color swatch match the basic calendar color).

`manifest.xml` needs some attention -- need to figure out what the trigger really is for calling this.